refactor: Rename dtos for taking authentation data and sending token
